chore: move SelectionCard to general components folder
