+++ src/client/v2/src/components/AddAccountModal/steps/SelectBankStep.tsx
@@ -4,8 +4,8 @@ import { IconSearch } from '@tabler/icons-react';
 import { Loader, Stack, Text, TextInput } from '@mantine/core';
 import { useQuery } from '@apollo/client/react';
 import { GetBankingInstitutions } from '@graphql-queries/GetBankingInstitutions';
+import { SelectionCard } from '../../SelectionCard/SelectionCard';
 import classes from '../AddAccountModal.module.css';
-import { SelectionCard } from './SelectionCard';
 
 const COLUMNS = 2;
 // Row height: 1px top border + 12px padding + 38px avatar (md) + 12px padding + 1px bottom border = 64px
